Implement lower extremities fields with 5/355 colimator angle
Add configuration option to decide whether to use 90 degrees or 5/355 degrees for the collimator angle of the lower extremities fileds. Default is 5/355.

Split Isocenters class into lower- and upper-specific classes.
