+++ TMIAutomation/Helper/Configuration/ConfigOptOptions.cs
@@ -32,6 +32,7 @@ namespace TMIAutomation
         public static string DosePerFraction => optSettings["DosePerFraction"];
         public static string NumberOfFractions => optSettings["NumberOfFractions"];
         public static string TreatmentMachine => optSettings["TreatmentMachine"];
+        public static string LowerExtremitiesCollimator => optSettings.ContainsKey("LowerExtremitiesCollimator") ? optSettings["LowerExtremitiesCollimator"] : string.Empty;
         public static bool BaseDosePlanning =>
 #if ESAPI15
                 optSettings.ContainsKey("BaseDosePlanning") && optSettings["BaseDosePlanning"] == "Yes";
@@ -39,6 +40,5 @@ namespace TMIAutomation
                 true; // Always true for ESAPI16 and ESAPI18
         public static bool AutoPlanLowerExtremities => optSettings["AutoPlanLowerExtremities"] == "Yes";
 #endif
-
     }
 }
